server folder in models (#23316)
